+++ tiger-rbel/src/main/java/de/gematik/rbellogger/converter/RbelCetpConverter.java
@@ -45,6 +45,7 @@ public class RbelCetpConverter implements RbelConverterPlugin {
 
     final RbelCetpFacet cetpFacet =
         RbelCetpFacet.builder()
+            .menuInfoString("CETP")
             .messageLength(RbelElement.wrap(messageLengthBytes, targetElement, messageLength))
             .body(converter.convertElement(messageBody, targetElement))
             .build();