+++ tiger-test-lib/src/main/java/de/gematik/test/tiger/glue/AnnotationParser.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2024 gematik GmbH
+ * 
+ * Licensed under the Apache License, Version 2.0 (the License);
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an 'AS IS' BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package de.gematik.test.tiger.glue;
 
 import com.github.javaparser.StaticJavaParser;
@@ -30,9 +46,14 @@ public class AnnotationParser {
   }
 
   public static void main(String[] args) throws IOException {
-    AnnotationParser annotationParser =
-        new AnnotationParser(
-            "tiger-test-lib/src/main/java/de/gematik/test/tiger/glue/", "./doc/user_manual/");
+    File f = new File(".");
+    String folder = "src/main/java/de/gematik/test/tiger/glue/";
+    String docFolder = "../doc/user_manual/";
+    if (!f.getAbsolutePath().endsWith("tiger-test-lib/.")) {
+      folder = "tiger-test-lib/" + folder;
+      docFolder = "./doc/user_manual/";
+    }
+    AnnotationParser annotationParser = new AnnotationParser(folder, docFolder);
     annotationParser.extractJavaDocsToAdoc(args);
   }