+++ tiger-testenv-mgr/src/frontend/src/types/rbel/MessageMetaDataDto.ts
@@ -18,6 +18,7 @@ export default class MessageMetaDataDto {
   uuid: string = '';
   path: string = '';
   method: string = '';
+  menuInfoString: string = '';
   responseCode: number = -1;
   recipient: string = '';
   sender: string = '';