+++ tiger-testenv-mgr/src/test/java/de/gematik/test/tiger/testenvmgr/TigerProxyMeshTest.java
@@ -151,6 +151,13 @@ class TigerProxyMeshTest extends AbstractTestTigerTestEnvMgr {
         .extractChildWithPath("$.receiver.bundledServerName")
         .hasStringContentEqualTo("httpbin");
 
+    assertThat(envMgr.getLocalTigerProxyOrFail().getRbelLogger().getMessageList().get(0))
+        .extractChildWithPath("$.receiver.port")
+        .hasStringContentEqualTo(TigerGlobalConfiguration.readString("free.port.0"));
+    assertThat(envMgr.getLocalTigerProxyOrFail().getRbelLogger().getMessageList().get(1))
+        .extractChildWithPath("$.sender.port")
+        .hasStringContentEqualTo(TigerGlobalConfiguration.readString("free.port.0"));
+
     waitShortTime();
   }
 
@@ -321,6 +328,8 @@ class TigerProxyMeshTest extends AbstractTestTigerTestEnvMgr {
   @TigerTest(
       tigerYaml =
           """
+         lib:
+             trafficVisualization: true
          tigerProxy:
            skipTrafficEndpointsSubscription: true
            adminPort: ${free.port.36}
@@ -341,7 +350,7 @@ class TigerProxyMeshTest extends AbstractTestTigerTestEnvMgr {
 
     try (Socket clientSocket =
         new Socket(
-            "127.0.0.1",
+            "localhost",
             Integer.parseInt(TigerGlobalConfiguration.resolvePlaceholders("${free.port.35}")))) {
 
       clientSocket.setSoTimeout(1);
@@ -366,6 +375,10 @@ class TigerProxyMeshTest extends AbstractTestTigerTestEnvMgr {
           .extractChildWithPath("$.sender")
           .asString()
           .isIn("localhost:" + senderPort, "127.0.0.1:" + senderPort);
+
+      assertThat(message)
+          .extractChildWithPath("$.receiver.bundledServerName")
+          .hasStringContentEqualTo("reverseHostname");
     }
     waitShortTime();
   }